refactor(pdf-service): move index from /services/pdf/attachment to /domains/shared/pdf/attachment
