+++ app/domains/beratungshilfe/services/pdf/pdfForm/E_unterhalt.ts
@@ -1,5 +1,5 @@
+import { SEE_IN_ATTACHMENT_DESCRIPTION } from "~/domains/shared/pdf/attachment";
 import { getFillUnterhalt } from "~/domains/shared/pdf/unterhaltHelpers";
-import { SEE_IN_ATTACHMENT_DESCRIPTION } from "~/services/pdf/attachment";
 import type { BerHPdfFillFunction } from "..";
 
 export const ATTACHMENT_DESCRIPTION_SECTION_E = "Feld E: Unterhaltszahlungen";
