+++ app/domains/prozesskostenhilfe/services/pdf/pdfForm/E_bruttoEinnahmen/__test__/bruttoEinnahmen_partner.test.ts
@@ -3,7 +3,7 @@ import type { ProzesskostenhilfePDF } from "data/pdf/prozesskostenhilfe/prozessk
 import { getProzesskostenhilfeParameters } from "data/pdf/prozesskostenhilfe/prozesskostenhilfe.generated";
 import { createFinancialEntry } from "tests/fixtures/prozesskostenhilfeFormularData";
 import { CheckboxValue } from "~/components/inputs/Checkbox";
-import { SEE_IN_ATTACHMENT_DESCRIPTION } from "~/services/pdf/attachment";
+import { SEE_IN_ATTACHMENT_DESCRIPTION } from "~/domains/shared/pdf/attachment";
 import * as fillOutFunctions from "~/services/pdf/fillOutFunction";
 import {
   fillAndereLeistungenPartner,
