+++ app/domains/prozesskostenhilfe/services/pdf/pdfForm/E_bruttoEinnahmen/bruttoEinnahmen_partner.ts
@@ -2,7 +2,7 @@ import { partnerEinkuenfteGuards as guards } from "~/domains/prozesskostenhilfe/
 import {
   type AttachmentEntries,
   SEE_IN_ATTACHMENT_DESCRIPTION,
-} from "~/services/pdf/attachment";
+} from "~/domains/shared/pdf/attachment";
 import { pdfFillReducer } from "~/services/pdf/fillOutFunction";
 import {
   nettoString,
