+++ app/domains/prozesskostenhilfe/services/pdf/pdfForm/G_eigentum.ts
@@ -1,4 +1,8 @@
 import {
+  SEE_IN_ATTACHMENT_DESCRIPTION,
+  type AttachmentEntries,
+} from "~/domains/shared/pdf/attachment";
+import {
   eigentuemerMapping,
   attachBankkontenToAnhang,
   grundeigentumArtMapping,
@@ -10,10 +14,6 @@ import {
   fillSingleWertsache,
   attachGeldanlagenToAnhang,
 } from "~/domains/shared/pdf/eigentumHelpers";
-import {
-  SEE_IN_ATTACHMENT_DESCRIPTION,
-  type AttachmentEntries,
-} from "~/services/pdf/attachment";
 import { pdfFillReducer } from "~/services/pdf/fillOutFunction";
 import { arrayIsNonEmpty } from "~/util/array";
 import type { PkhPdfFillFunction } from "..";
