+++ app/domains/shared/pdf/attachment/index.ts
